combine into one file
(cherry picked from commit 8cf0b33bb2b15ffb4440a8832464eadad2fa1ef6)
