--- tests/sentry/db/models/fields/bitfield/forms.py
@@ -1,11 +1,0 @@
-from __future__ import absolute_import
-
-from django import forms
-
-from .models import BitFieldTestModel
-
-
-class BitFieldTestModelForm(forms.ModelForm):
-    class Meta:
-        model = BitFieldTestModel
-        exclude = tuple()
